requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Quantum Computing
qiskit==1.0.0
qiskit-aer==0.13.3
qiskit-ibm-runtime==0.18.0
scipy==1.12.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.1
rich==13.7.0
requests==2.31.0

# Storage
ipfshttpclient==0.8.0a2

# Monitoring
prometheus-client==0.19.0

# Data validation
pydantic==2.5.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Binary compilation
pyinstaller==6.3.0
